+++ tests/HtmlRendererTest.php
@@ -8,14 +8,18 @@
 
 namespace SebLucas\Cops\Tests;
 
+use SebLucas\Cops\Output\DotPHPTemplate;
 use SebLucas\Cops\Output\HtmlRenderer;
+use SebLucas\Cops\Output\TwigTemplate;
 
 require_once dirname(__DIR__) . '/config/test.php';
 use PHPUnit\Framework\TestCase;
 use SebLucas\Cops\Calibre\Database;
 use SebLucas\Cops\Input\Config;
 use SebLucas\Cops\Input\Request;
+use SebLucas\Cops\Output\JsonRenderer;
 use SebLucas\Cops\Pages\PageId;
+use DOMDocument;
 
 class HtmlRendererTest extends TestCase
 {
@@ -26,7 +30,77 @@ class HtmlRendererTest extends TestCase
         $_GET = [];
     }
 
-    public function testHtmlRenderer(): void
+    /**
+     * The function for the head of the HTML catalog
+     * @param mixed $templateName
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\DataProvider('providerDotTemplate')]
+    public function testDotHeader($templateName)
+    {
+        $_SERVER["HTTP_USER_AGENT"] = "Firefox";
+        $_COOKIE["template"] = $templateName;
+        $request = new Request();
+        $html = new HtmlRenderer();
+        $data = $html->getTemplateData($request);
+        $template = new DotPHPTemplate($request);
+        $tpl = $template->getDotTemplate(dirname(__DIR__) . '/templates/' . $templateName . '/file.html');
+
+        $head = $tpl($data);
+        $this->assertStringContainsString("<head>", $head);
+        $this->assertStringContainsString("</head>", $head);
+
+        unset($_COOKIE["template"]);
+    }
+
+    /**
+     * FALSE is returned if the create_function failed (meaning there was a syntax error)
+     * @param mixed $templateName
+     * @return void
+     */
+    #[\PHPUnit\Framework\Attributes\DataProvider('providerDotTemplate')]
+    public function testDotServerSide($templateName)
+    {
+        $_COOKIE["template"] = $templateName;
+        $request = new Request();
+        $template = new DotPHPTemplate($request);
+        $this->assertNull($template->serverSide(null, $templateName));
+
+        $json = new JsonRenderer();
+        $data = $json->getJson($request, true);
+        $output = $template->serverSide($data, $templateName);
+
+        $old = libxml_use_internal_errors(true);
+        $html = new DOMDocument();
+        $html->loadHTML("<html><head></head><body>" . $output . "</body></html>");
+        $errors = libxml_get_errors();
+        foreach ($errors as $error) {
+            // ignore invalid tags from HTML5 which libxml doesn't know about
+            if ($error->code == 801) {
+                continue;
+            }
+            $this->fail("Error parsing output for server-side rendering of template " . $templateName . "\n" . $error->message . "\n" . $output);
+        }
+        libxml_clear_errors();
+        libxml_use_internal_errors($old);
+
+        unset($_COOKIE['template']);
+    }
+
+    /**
+     * Summary of providerDotTemplate
+     * @return array<mixed>
+     */
+    public static function providerDotTemplate()
+    {
+        return [
+            ["bootstrap2"],
+            ["bootstrap"],
+            ["default"],
+        ];
+    }
+
+    public function testRenderDot(): void
     {
         $page = PageId::ALL_RECENT_BOOKS;
         $request = Request::build(['page' => $page]);
@@ -42,7 +116,7 @@ class HtmlRendererTest extends TestCase
         $this->assertStringContainsString($expected, $output);
     }
 
-    public function testHtmlRendererServerSide(): void
+    public function testRenderDotServerSide(): void
     {
         $_SERVER['HTTP_USER_AGENT'] = "Kindle/1.0";
 
@@ -58,7 +132,41 @@ class HtmlRendererTest extends TestCase
         unset($_SERVER['HTTP_USER_AGENT']);
     }
 
-    public function testHtmlRendererTwig(): void
+    /**
+     * Summary of testTwigServerSide
+     * @param mixed $templateName
+     * @return void
+     */
+    public function testTwigServerSide($templateName = 'twigged')
+    {
+        $_COOKIE["template"] = $templateName;
+        $request = new Request();
+        $template = new TwigTemplate($request);
+        $twig = $template->getTwigEnvironment('templates/' . $templateName);
+        $this->assertNull($template->serverSide($twig, null, $templateName));
+
+        $json = new JsonRenderer();
+        $data = $json->getJson($request, true);
+        $output = $template->serverSide($twig, $data, $templateName);
+
+        $old = libxml_use_internal_errors(true);
+        $html = new DOMDocument();
+        $html->loadHTML("<html><head></head><body>" . $output . "</body></html>");
+        $errors = libxml_get_errors();
+        foreach ($errors as $error) {
+            // ignore invalid tags from HTML5 which libxml doesn't know about
+            if ($error->code == 801) {
+                continue;
+            }
+            $this->fail("Error parsing output for server-side rendering of template " . $templateName . "\n" . $error->message . "\n" . $output);
+        }
+        libxml_clear_errors();
+        libxml_use_internal_errors($old);
+
+        unset($_COOKIE['template']);
+    }
+
+    public function testRenderTwig(): void
     {
         $_COOKIE['template'] = "twigged";
 
@@ -78,7 +186,7 @@ class HtmlRendererTest extends TestCase
         unset($_COOKIE['template']);
     }
 
-    public function testHtmlRendererTwigServerSide(): void
+    public function testRenderTwigServerSide(): void
     {
         $_SERVER['HTTP_USER_AGENT'] = "Kindle/1.0";
         $_COOKIE['template'] = "twigged";